core/design/requirements.txt
# DEAH Design Agents — Python dependencies
# Install with: pip install -r requirements.txt

# API server
fastapi>=0.111.0
uvicorn[standard]>=0.29.0

# Streamlit frontend
streamlit>=1.35.0

# Claude / Anthropic SDK (direct API)
anthropic>=0.30.0

# Claude Code Agent SDK (architecture agent uses claude-code-sdk provider)
claude-agent-sdk
anyio>=4.0.0

# Environment variables
python-dotenv>=1.0.0

# Data processing (data model agent uses pandas for CSV handling)
pandas>=2.0.0

# HTTP client (requirements agent Jira calls)
httpx>=0.27.0
requests
# SQLAlchemy — used by agent_output_metadata (metadata DB reads/writes)
sqlalchemy>=2.0.0

# MySQL driver for SQLAlchemy (metadata DB is MySQL)
pymysql>=1.1.0

# Database connectors (only needed if connecting to live source databases)
# Uncomment the ones you need:
# mysql-connector-python>=8.0.0
# psycopg2-binary>=2.9.0
# pyodbc>=4.0.0
# python-oracledb>=2.0.0